+++ index.html
@@ -8,12 +8,12 @@
 <body>
 
     <h1>Todo list</h1>
-    <input type="text" id="taskInput" placeholder="Write something you needed to do.">
-    <button id="createBtn">Create</button>
-    <button onclick="test()">TestSession</button>
-    <button onclick="clearSession()">clear</button><br>
-    <div id="container"></div>
+    <input type="text" class="taskInput" placeholder="Write something you needed to do.">
+    <button class="createBtn">Create</button>
+    <button class="checkSession">check</button>
+    <button class="clearSession">clear</button><br>
+    <div class="container"></div>
 
-    <script src="script.js"></script>
+    <script type="module" src="app.js"></script>
 </body>
 </html>
