+++ modules/components.js
@@ -1,0 +1,2 @@
+export const taskInput = document.querySelector(".taskInput");
+export const container = document.querySelector(".container");
